Extract VersioningResourceResolver as interface
