+++ platform/base/core/src/main/java/com/peregrine/versions/VersioningRequestFilter.java
@@ -1,6 +1,8 @@
 package com.peregrine.versions;
 
+import org.apache.commons.lang3.StringUtils;
 import org.apache.sling.api.SlingHttpServletRequest;
+import org.apache.sling.api.resource.PersistenceException;
 import org.apache.sling.api.resource.Resource;
 import org.apache.sling.api.resource.ResourceResolver;
 import org.apache.sling.api.servlets.ServletResolver;
@@ -10,7 +12,6 @@ import org.osgi.framework.Constants;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
-import javax.jcr.RepositoryException;
 import javax.servlet.*;
 import javax.servlet.http.Cookie;
 import java.io.IOException;
@@ -19,6 +20,7 @@ import java.lang.reflect.Method;
 import java.util.Optional;
 
 import static com.peregrine.commons.Chars.EQ;
+import static com.peregrine.versions.VersioningResourceResolver.LABEL_PROPERTY;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
 
 @Component(
@@ -62,9 +64,12 @@ public final class VersioningRequestFilter implements Filter {
     }
 
     private static String extractLabel(final SlingHttpServletRequest request) {
-        return Optional.ofNullable(request.getCookie(VersioningResourceResolver.LABEL_PROPERTY))
+        return Optional.ofNullable(request.getCookie(LABEL_PROPERTY))
                 .map(Cookie::getValue)
-                .orElse(null);
+                .filter(StringUtils::isNotBlank)
+                .orElseGet(() -> Optional.ofNullable(request.getAttribute(LABEL_PROPERTY))
+                        .map(Object::toString)
+                        .orElse(null));
     }
 
     private void switchResolver(final SlingHttpServletRequest request, final String label) throws ServletException {
@@ -74,12 +79,12 @@ public final class VersioningRequestFilter implements Filter {
             final Object requestData = getRequestData.invoke(request);
             final Class<?> requestDataClass = requestData.getClass();
             final Method initResource = requestDataClass.getMethod("initResource", ResourceResolver.class);
-            final ResourceResolver resolver = new VersioningResourceResolver(oldResolver, label);
+            final ResourceResolver resolver = new VersioningResourceResolverImpl(oldResolver, label);
             final Object resource = initResource.invoke(requestData, resolver);
             final Method initServlet = requestDataClass
                     .getMethod("initServlet", Resource.class, ServletResolver.class);
             initServlet.invoke(requestData, resource, servletResolver);
-        } catch (final NoSuchMethodException | InvocationTargetException | IllegalAccessException | RepositoryException e) {
+        } catch (final NoSuchMethodException | InvocationTargetException | IllegalAccessException | PersistenceException e) {
             throw new ServletException("Error switching ResourceResolver");
         }
     }
